+++ commands/fun/rr.js
@@ -71,7 +71,7 @@ module.exports.run = async (msg, Flags, Parameters, bot, ret)=>{
 
 module.exports.config = {
     name:'rr',
-    desc:'Playing game of chanse.',
+    desc:'Playing game of chance.',
     permissions:'000001',
     parameters:[],
     flags:{},