+++ commands/reputation/removerep.js
@@ -11,10 +11,11 @@ module.exports.run = (msg, Flags, Parameters, bot, ret)=>{
         ret.message = Math.round(Math.random() * 1000) == 0 ? 'Dont you try again.' : 'You can\'t give reputation to yourself';
         return ret;
     }
-    const lastReputationInteraction = Date.parse(userData.getUserConfigKey(giver.id, 'lastReputationInteraction')) || 10000000000;
+    const lris = userData.getUserConfigKey(giver.id, 'lastReputationInteraction');
+    const lastReputationInteraction = lris != '' ? Date.parse(lris) : 100000000000;
     // Stack overflow (STOLEN CODE HERE)
     const diffMs = now.getTime() - lastReputationInteraction;
-    const diffMins = Math.round(((diffMs % 86400000) % 3600000) / 60000); // minutes
+    const diffMins = Math.abs(Math.round(((diffMs % 86400000) % 3600000) / 60000)); // minutes
 
     if(diffMins < 30) {
         const embed = new discord.MessageEmbed();